requirements.txt
pyserial>=2.6
git+https://github.com/FrederikBjorne/python-observer
PyQT5
